feat(cell-cards): tabbed marker gene tables
