--- frontend/src/views/CellCards/components/CellCard/components/CanonicalMarkerGeneTable/style.ts
@@ -1,6 +1,0 @@
-import styled from "@emotion/styled";
-
-export const PublicationLinkWrapper = styled.div`
-  display: flex;
-  column-gap: 2px;
-`;
