--- frontend/src/views/CellCards/components/CellCard/components/EnrichedGenesTable/style.ts
@@ -1,8 +1,0 @@
-import styled from "@emotion/styled";
-
-export const TableTitleOuterWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-`;
