+++ frontend/src/views/CellCards/components/CellCard/components/SourceDataTable/index.tsx
@@ -6,6 +6,7 @@ import {
   TableUnavailableContainer,
   TableUnavailableHeader,
   TableUnavailableDescription,
+  StyledDivider,
 } from "../common/style";
 import Table from "../common/Table";
 import Link from "../common/Link";
@@ -141,6 +142,7 @@ const SourceDataTable = ({ cellTypeId }: Props) => {
       <TableTitleWrapper>
         <TableTitle>Data</TableTitle>
       </TableTitleWrapper>
+      <StyledDivider />
       {tableRows.length > 0 ? (
         <Table<TableRow> columns={tableColumns} rows={tableRows} />
       ) : (
